+++ History.md
@@ -1,6 +1,9 @@
 <!-- HISTORY -->
 ## History
 
+* 0.41.5 - 29.03.2026
+  * Google Chrome Ubuntu fix.
+  * Sudo delegation fix on Ubuntu.
 * 0.41.4 - 24.03.2026
   * Dependencies update.
 * 0.41.3 - 07.03.2026